feat: set description for VIP 347
